% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decor.R
\name{cpp_decorations}
\alias{cpp_decorations}
\title{Decorations in a C++ file}
\usage{
cpp_decorations(pkg = ".", files = cpp_files(pkg = pkg))
}
\arguments{
\item{pkg}{A package, see `devtools::as.package()`}

\item{files}{C++ files}
}
\description{
Decorations in a C++ file
}
